requirements.txt
fastapi
uvicorn
stripe
python-dotenv
python-multipart
httpx
